+++ aiosmtpd/docs/NEWS.rst
@@ -6,6 +6,7 @@
 ==================
 * Fix support for both IPv4 and IPv6 based on the --listen option.  Given by
   Jason Coombs.  (Closes #3)
+* Flake8 clean.
 
 1.0a1 (2015-10-19)
 ==================